+++ apps/backend/backend_commit.c
@@ -349,6 +349,9 @@ startup_commit(clicon_handle  h,
      /* 8. Call plugin transaction commit callbacks */
      if (plugin_transaction_commit(h, td) < 0)
 	 goto done;
+     /* After commit, make a post-commit call (sure that all plugins have committed) */
+     if (plugin_transaction_commit_done(h, td) < 0)
+	 goto done;
      /* Clear cached trees from default values and marking */
      if (xmldb_get0_clear(h, td->td_target) < 0)
 	 goto done;
@@ -537,6 +540,9 @@ candidate_commit(clicon_handle h,
 
      /* 7. Call plugin transaction commit callbacks */
      if (plugin_transaction_commit(h, td) < 0)
+	 goto done;
+     /* After commit, make a post-commit call (sure that all plugins have committed) */
+     if (plugin_transaction_commit_done(h, td) < 0)
 	 goto done;
      
      /* Clear cached trees from default values and marking */